[FIX] mail_activity: prevent inconsistent team value
Fixes 'Assigned user is not member of the team' when an activity is created
in the backend for a user that is not in the same teams as the user creating
the activity.

An example of such an activity is the one created for tax closing moves in
enterprises' account_reports.